+++ cli/src/lib/paths.js
@@ -75,10 +75,18 @@ module.exports = (cwd = process.cwd()) => {
 
         buildOutput: path.join(base, './build'),
         buildAppOutput: path.join(base, './build/app'),
+        buildAppManifest: path.join(base, './build/app/manifest.webapp'),
+        buildAppConfigJson: path.join(base, './build/app/d2.config.json'),
+        buildAppBundleFile: '{name}-{version}.zip',
+        buildAppBundleOutput: path.join(base, 'build', 'bundle'),
         buildAppBundle: path.join(
             base,
-            './build/bundle/dhis2-{{name}}-{{version}}.zip'
+            'build',
+            'bundle',
+            '{name}-{version}.zip'
         ),
+        buildLibBundleFile: '{name}-{version}.zip',
+        buildLibBundleOutput: path.join(base),
     }
 
     reporter.debug('PATHS', paths)